+++ dialbb/builtin_blocks/chatgpt/chatgpt.py
@@ -25,14 +25,21 @@ __copyright__ = 'C4A Research Institute, Inc.'
 import os
 import sys
 import traceback
+import datetime
 from typing import Dict, Any, List, Union, Tuple
 import openai
 from dialbb.abstract_block import AbstractBlock
 from dialbb.util.error_handlers import abort_during_building
+import re
 
 DIALOGUE_HISTORY_OLD_TAG: str = '@dialogue_history'
 DIALOGUE_HISTORY_TAG: str = '{dialogue_history}'
+CURRENT_TIME_TAG: str = '{current_time}'
 DEFAULT_GPT_MODEL: str = "gpt-4o-mini"
+
+#  [[[....{tag1}....{tag2}....]]]
+REMAINING_TAGS_PATTERN = re.compile( r"\[\[\[(?s)(?=.*\{[A-Za-z0-9_]+\})(?:[^\{\]]|\{[A-Za-z0-9_]+\})*\]\]\]",
+                                     re.DOTALL)
 
 
 class ChatGPT(AbstractBlock):
@@ -130,6 +137,25 @@ class ChatGPT(AbstractBlock):
         system_utterance: str = chat_completion.choices[0].message.content
         return system_utterance
 
+    @staticmethod
+    def get_current_time_string(language: str) -> str:
+        """
+        generates string to represent current string.
+        :param language:
+        :return:
+        """
+
+        now = datetime.datetime.now()
+        if language == 'ja':
+            weekdays = ["月", "火", "水", "木", "金", "土", "日"]
+            date_str = now.strftime("%Y年%m月%d日")
+            time_str = now.strftime("%H時%M分%S秒")
+            weekday_str = weekdays[now.weekday()]
+            result: str = f"{date_str}（{weekday_str}） {time_str}"
+        else:
+            result = now.strftime("%A, %B %d, %Y %I:%M:%S %p")
+        return result
+
     def generate_system_utterance(self, dialogue_history: List[Dict[str, str]],
                                   session_id: str, user_id: str,
                                   aux_data: Dict[str, Any]) -> Tuple[str, Dict[str, Any], bool]:
@@ -144,8 +170,18 @@ class ChatGPT(AbstractBlock):
         :return: a tuple of system utterance string, aux_data as is, and final flag (always False)
         """
 
+        language = self.config.get("langauge", 'en')
+        prompt = self._prompt_template
+        prompt = prompt.replace(CURRENT_TIME_TAG, self.get_current_time_string(language))  # {current_time}
+        if aux_data:
+            for aux_data_key, aux_data_value in aux_data.items():  # aux_data values replace their place holders
+                prompt = prompt.replace("{" + aux_data_key + "}", str(aux_data_value))
+        prompt = REMAINING_TAGS_PATTERN.sub("", prompt)  # remove remaining tags enclosed by [[[ .... ]]]]
+        prompt = prompt.replace('[[[', "")  # remove remaining brackets
+        prompt = prompt.replace(']]]', "")
+
         messages = []
-        messages.append({'role': "system", "content": self._prompt_template})
+        messages.append({'role': "system", "content": prompt})
 
         for turn in dialogue_history:
             if turn["speaker"] == 'user':