+++ Changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+
+## 1.2.0 (xxxx.xx.xx)
+
+- ChatGPT Dialogue Embedded Block
+
+  - Changed to allow inserting keys from aux_data and the current date/time into prompt template placeholders.
+
+- STN Manager Embedded Block
+
+  - Changed to allow inserting keys from aux_data into prompt template placeholders.
+	
 ## 1.1.0 (2025.11.8)
 
 - No-code tool
@@ -9,8 +20,8 @@
  - The appearance of the scenario editor enhanced
 
  - The default applications changed
-	
- - The chat window implemented	
+
+ - The chat window implemented 
 
 - Embedded block for ChatGPT Dialogue:
 
@@ -183,14 +194,23 @@
 ## 0.2.0 (2022.12.1)
 
 - Interface changes in AbstractBlock (not backward compatible)
+
   - Changes to the arguments of the process method
-- Set default language to Japanese
+	
+  - Set default language to Japanese
+
 - Changes to the STN Manager block
+
   - Exporting scenario graphs
+
   - Introduction of preparation state
+
 - Enabled the use of Google Sheets
+
 - Changed the test scenario format (not backward compatible)
+
 - Column name changes for knowledge description for SNIPS Understander (not backward compatible)
+
 - Support for n-best outputs in SNIPS Understander
 
 ## 0.1 (2022.8.9)